# swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  AppCredentialsPayload:
    description: App ID+secret credentials
    example:
      id: Itaque magnam consequatur doloribus.
      secret: Inventore est.
    properties:
      id:
        description: The app ID
        example: Itaque magnam consequatur doloribus.
        type: string
      secret:
        description: The app secret
        example: Inventore est.
        type: string
    required:
    - id
    - secret
    title: AppCredentialsPayload
    type: object
  AppPayload:
    description: Payload for the client apps
    example:
      description: dquu7sxd1b
      domain: Mollitia et quasi esse voluptate.
      name: zzr28p88rb
    properties:
      description:
        description: Description of the app
        example: dquu7sxd1b
        maxLength: 300
        type: string
      domain:
        description: App domain
        example: Mollitia et quasi esse voluptate.
        type: string
      name:
        description: Name of the app
        example: zzr28p88rb
        maxLength: 50
        type: string
    required:
    - name
    title: AppPayload
    type: object
  apps:
    description: apps media type (default view)
    example:
      description: lx1y6tc2l6
      domain: Quae earum.
      id: Possimus vel.
      name: f0iuv3mp0p
      owner: In rerum.
      registeredAt: 2717061749445211706
    properties:
      description:
        description: Description of the app
        example: lx1y6tc2l6
        maxLength: 300
        type: string
      domain:
        description: App domain
        example: Quae earum.
        type: string
      id:
        description: Unique app ID
        example: Possimus vel.
        type: string
      name:
        description: Name of the app
        example: f0iuv3mp0p
        maxLength: 50
        type: string
      owner:
        description: User ID
        example: In rerum.
        type: string
      registeredAt:
        description: Time when app is registered
        example: 2717061749445211706
        format: int64
        type: integer
    required:
    - id
    - name
    - description
    - domain
    - owner
    - registeredAt
    title: 'Mediatype identifier: application/vnd.goa.apps+json; view=default'
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1458609066
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1458609066
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
  reg-apps:
    description: reg-apps media type (default view)
    example:
      id: Impedit ea laborum vitae.
      secret: Eum qui sed placeat est.
    properties:
      id:
        description: App ID
        example: Impedit ea laborum vitae.
        type: string
      secret:
        description: Client secret
        example: Eum qui sed placeat est.
        type: string
    required:
    - id
    - secret
    title: 'Mediatype identifier: application/vnd.goa.reg.apps+json; view=default'
    type: object
host: localhost:8080
info:
  description: A service that provides basic access to the applications management
  title: The apps management microservice
  version: "1.0"
paths:
  /:
    post:
      description: Register new app
      operationId: apps#registerApp
      parameters:
      - description: Payload for the client apps
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AppPayload'
      produces:
      - application/vnd.goa.error
      - application/vnd.goa.reg.apps+json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/reg-apps'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: registerApp apps
      tags:
      - apps
  /{appId}:
    delete:
      description: Delete an app
      operationId: apps#deleteApp
      parameters:
      - in: path
        name: appId
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: deleteApp apps
      tags:
      - apps
    get:
      description: Get app by id
      operationId: apps#get
      parameters:
      - description: App ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/vnd.goa.apps+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apps'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: get apps
      tags:
      - apps
    put:
      description: Register new app
      operationId: apps#updateApp
      parameters:
      - in: path
        name: appId
        required: true
        type: string
      - description: Payload for the client apps
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AppPayload'
      produces:
      - application/vnd.goa.apps+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apps'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: updateApp apps
      tags:
      - apps
  /{appId}/regenerate-secret:
    put:
      description: Regenerate client secret
      operationId: apps#regenerateClientSecret
      parameters:
      - in: path
        name: appId
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: regenerateClientSecret apps
      tags:
      - apps
  /my:
    get:
      description: Get all user's apps
      operationId: apps#getMyApps
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: getMyApps apps
      tags:
      - apps
  /swagger-ui/{filepath}:
    get:
      operationId: swagger#/swagger-ui/*filepath
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: Download swagger-ui/dist
  /swagger.json:
    get:
      operationId: swagger#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
      summary: Download swagger/swagger.json
  /users/{userId}/all:
    get:
      description: Get app by id
      operationId: apps#getUserApps
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: getUserApps apps
      tags:
      - apps
  /verify:
    post:
      description: Verify an application by its ID and secret
      operationId: apps#verifyApp
      parameters:
      - description: App ID+secret credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AppCredentialsPayload'
      produces:
      - application/vnd.goa.apps+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apps'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: verifyApp apps
      tags:
      - apps
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
responses:
  OK:
    description: OK
schemes:
- http
swagger: "2.0"